backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
pyyaml==6.0.1
pydantic==2.6.3
uvicorn==0.27.1
